Let the add-repository flow outlive the picker dialog
Follow-up to the Lite onboarding PR (#15365), closing out problems found
in review:

- The button owned the whole pick -> add -> navigate flow, and the
  project picker closed itself before picking — unmounting the button
  and its in-flight flow with it
- useAddLocalRepository (own file, so fast refresh stays happy) is
  hosted by the page, which outlives the dialog; AddProjectButton is
  now purely presentational (isPending/onClick)
- failureMessage is total over AddProjectFailure: the success arms
  leave the switch instead of returning ""
- set -euo pipefail in the project-with-named-branch fixture

Verified with the onboarding and project-picker e2e journeys.
