+++ apps/lite/ui/src/routes/index.tsx
@@ -2,6 +2,7 @@ import { createRoute, redirect } from "@tanstack/react-router";
 import { LiteTestId } from "@gitbutler/ui/utils/testIds";
 import type { FC } from "react";
 import { AddProjectButton } from "#ui/components/AddProjectButton.tsx";
+import { useAddLocalRepository } from "#ui/components/useAddLocalRepository.ts";
 import { Route as rootRoute } from "#ui/routes/__root.tsx";
 import { readLastOpenedProject, readLastPlace } from "#ui/project.ts";
 import styles from "./IndexPage.module.css";
@@ -10,13 +11,21 @@ const parseLastSearch = (search: string): Record<string, string> =>
 	Object.fromEntries(new URLSearchParams(search));
 
 // oxlint-disable-next-line react/only-export-components -- False positive?
-const IndexPage: FC = () => (
-	<section className={styles.page} data-testid={LiteTestId.OnboardingPage}>
-		<h1>Welcome to GitButler Lite</h1>
-		<p>Add a local Git repository to get started.</p>
-		<AddProjectButton testId={LiteTestId.OnboardingAddLocalProjectButton} />
-	</section>
-);
+const IndexPage: FC = () => {
+	const { addLocalRepository, isPending } = useAddLocalRepository();
+
+	return (
+		<section className={styles.page} data-testid={LiteTestId.OnboardingPage}>
+			<h1>Welcome to GitButler Lite</h1>
+			<p>Add a local Git repository to get started.</p>
+			<AddProjectButton
+				testId={LiteTestId.OnboardingAddLocalProjectButton}
+				isPending={isPending}
+				onClick={() => void addLocalRepository()}
+			/>
+		</section>
+	);
+};
 
 export const Route = createRoute({
 	getParentRoute: () => rootRoute,
